+++ slices/remove_by_test.go
@@ -1,0 +1,113 @@
+package slices
+
+import (
+	"fmt"
+	"strconv"
+	"testing"
+
+	"github.com/golodash/godash/internal"
+)
+
+type TRemoveBy struct {
+	name      string
+	arg1      interface{}
+	wantSlice interface{}
+	wantRems  interface{}
+}
+
+var TRemoveByBenchs = []TRemoveBy{
+	{
+		name: "10",
+		arg1: []int{},
+	},
+	{
+		name: "100",
+		arg1: []int{},
+	},
+	{
+		name: "1000",
+		arg1: []int{},
+	},
+	{
+		name: "10000",
+		arg1: []int{},
+	},
+	{
+		name: "100000",
+		arg1: []int{},
+	},
+}
+
+func init() {
+	for i := 0; i < len(TRemoveByBenchs); i++ {
+		k, _ := strconv.Atoi(TRemoveByBenchs[i].name)
+		for j := 0; j < k/10; j++ {
+			TRemoveByBenchs[i].arg1 = append(TRemoveByBenchs[i].arg1.([]int), []int{1, 2, 3, 4, 5, 6, 7, 8, 9, 0}...)
+		}
+	}
+}
+
+func removeByFunctionTest(n interface{}) bool {
+	return n.(int)%2 != 0
+}
+
+func TestRemoveBy(t *testing.T) {
+	var tests = []TRemoveBy{
+		{
+			name:      "nil",
+			arg1:      nil,
+			wantSlice: nil,
+			wantRems:  nil,
+		},
+		{
+			name:      "empty",
+			arg1:      []int{},
+			wantSlice: []int{},
+			wantRems:  []int{},
+		},
+		{
+			name:      "default",
+			arg1:      []int{1, 2, 3, 4, 5, 6, 7, 8, 66, 44, 5, 6, 7, 99},
+			wantSlice: []int{2, 4, 6, 8, 66, 44, 6},
+			wantRems:  []int{99, 7, 5, 7, 5, 3, 1},
+		},
+		{
+			name:      "default1",
+			arg1:      []int{3, 4, 5, 6, 23, 34, 56, 68, 98},
+			wantSlice: []int{4, 6, 34, 56, 68, 98},
+			wantRems:  []int{23, 5, 3},
+		},
+	}
+
+	for _, sample := range tests {
+		t.Run(sample.name, func(t *testing.T) {
+			gotSlice, gotRems, err := RemoveBy(sample.arg1, removeByFunctionTest)
+			if err != nil {
+				if sample.wantSlice != nil && sample.wantRems != nil {
+					t.Errorf("gotSlice = %v, gotRem = %v, wantSlice = %v, wantRems = %v, err = %v", gotSlice, gotRems, sample.wantSlice, sample.wantRems, err)
+				}
+				return
+			}
+
+			if ok, _ := internal.Same(gotSlice, sample.wantSlice); !ok {
+				t.Errorf("gotSlice = %v, gotRem = %v, wantSlice = %v, wantRems = %v, err = %v", gotSlice, gotRems, sample.wantSlice, sample.wantRems, err)
+				return
+			}
+
+			if ok, _ := internal.Same(gotRems, sample.wantRems); !ok {
+				t.Errorf("gotSlice = %v, gotRem = %v, wantSlice = %v, wantRems = %v, err = %v", gotSlice, gotRems, sample.wantSlice, sample.wantRems, err)
+				return
+			}
+		})
+	}
+}
+
+func BenchmarkRemoveBy(b *testing.B) {
+	for _, sample := range TRemoveByBenchs {
+		b.Run(fmt.Sprintf("input_size_%s", sample.name), func(b *testing.B) {
+			for i := 0; i < b.N; i++ {
+				RemoveBy(sample.arg1, removeByFunctionTest)
+			}
+		})
+	}
+}